requirements.txt
tfx==1.2.0
absl-py==0.12.0
tensorflow~=2.5.1
pyyaml==5.4.1
pandas~=1.3.3